ecom-microservices/book-service/requirements.txt
Django==3.2.20
djongo==1.3.6
pymongo==3.12.3
dnspython==2.2.1
gunicorn==20.1.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
django-filter==21.1
